Address updater signing review feedback
