+++ src/renderer/src/components/UpdateCard.test.ts
@@ -1,21 +1,9 @@
-// @vitest-environment happy-dom
-/* eslint-disable max-lines -- Why: the update-card behavior is easiest to verify as one
-   lifecycle-oriented suite because the store caching and visibility rules interact directly. */
 import { createStore, type StoreApi } from 'zustand/vanilla'
-import { createElement } from 'react'
-import { act } from 'react'
-import { createRoot, type Root } from 'react-dom/client'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { getDefaultUIState } from '../../../shared/constants'
 import type { ChangelogData, UpdateStatus } from '../../../shared/types'
 import { createUISlice } from '../store/slices/ui'
-import { useAppStore } from '../store'
 import type { AppState } from '../store/types'
-import {
-  isHttp2ProtocolError,
-  shouldShowManualDownloadForUpdateError,
-  UpdateCard
-} from './UpdateCard'
 
 // ── Helpers ──────────────────────────────────────────────────────────
 
@@ -313,362 +301,6 @@ describe('hydratePersistedUI update fields', () => {
   })
 })
 
-// ── Visibility gate logic (pure function extraction) ─────────────────
-// Why: the UpdateCard component's render gates are the most critical
-// correctness surface — a wrong gate means users either miss updates or
-// get spammed with irrelevant cards. We test the gate logic as a pure
-// function to avoid needing a full React/DOM environment.
-
-type VisibilityInput = {
-  status: UpdateStatus
-  dismissedVersion: string | null
-  cachedVersion: string | null
-  hasStartedDownload: boolean
-  updateUserInitiatedCycle?: boolean
-}
-
-type VisibilityResult = 'hidden' | 'visible'
-
-/** Mirrors the visibility gates in UpdateCard's render path. */
-function computeVisibility(input: VisibilityInput): VisibilityResult {
-  const { status, dismissedVersion, cachedVersion, hasStartedDownload } = input
-  const isUserInitiated = 'userInitiated' in status && status.userInitiated
-  const updateUserInitiatedCycle = input.updateUserInitiatedCycle ?? false
-  const shouldShowDetailedErrorCard =
-    status.state === 'error' && (hasStartedDownload || cachedVersion !== null)
-
-  if (status.state === 'checking' && !isUserInitiated) {
-    return 'hidden'
-  }
-  if (status.state === 'not-available' && !isUserInitiated) {
-    return 'hidden'
-  }
-  if (status.state === 'idle') {
-    return 'hidden'
-  }
-  if (status.state === 'error' && !shouldShowDetailedErrorCard && !isUserInitiated) {
-    return 'hidden'
-  }
-
-  const effectiveVersion = 'version' in status ? status.version : cachedVersion
-  if (effectiveVersion && dismissedVersion === effectiveVersion && !updateUserInitiatedCycle) {
-    if (status.state !== 'downloading' && status.state !== 'error') {
-      return 'hidden'
-    }
-  }
-
-  return 'visible'
-}
-
-describe('UpdateCard visibility gates', () => {
-  it('hides on idle', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'idle' },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-
-  it('hides background checking (not user-initiated)', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'checking' },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-
-  it('shows user-initiated checking', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'checking', userInitiated: true },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('hides background not-available', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'not-available' },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-
-  it('shows user-initiated not-available (before auto-dismiss)', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'not-available', userInitiated: true },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('shows available update (simple mode)', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'available', version: '1.2.0', changelog: null },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('shows available update (rich mode)', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'available', version: '1.2.0', changelog: RICH_CHANGELOG },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('hides available when version is dismissed', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'available', version: '1.2.0', changelog: null },
-        dismissedVersion: '1.2.0',
-        cachedVersion: '1.2.0',
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-
-  it('shows dismissed available update when a lazy-mounted manual check cycle reaches available', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'available', version: '1.2.0', changelog: null },
-        dismissedVersion: '1.2.0',
-        cachedVersion: '1.2.0',
-        hasStartedDownload: false,
-        updateUserInitiatedCycle: true
-      })
-    ).toBe('visible')
-  })
-
-  it('shows downloading even when version is dismissed (user clicked Update after dismiss)', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'downloading', percent: 42, version: '1.2.0' },
-        dismissedVersion: '1.2.0',
-        cachedVersion: '1.2.0',
-        hasStartedDownload: true
-      })
-    ).toBe('visible')
-  })
-
-  it('hides downloaded when version is dismissed (Settings-initiated, not card)', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'downloaded', version: '1.2.0' },
-        dismissedVersion: '1.2.0',
-        cachedVersion: '1.2.0',
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-
-  it('hides background errors silently', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'error', message: 'network' },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-
-  it('shows user-initiated check errors', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'error', message: 'network', userInitiated: true },
-        dismissedVersion: null,
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('shows card-initiated download errors', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'error', message: 'ENOSPC' },
-        dismissedVersion: null,
-        cachedVersion: '1.2.0',
-        hasStartedDownload: true
-      })
-    ).toBe('visible')
-  })
-
-  it('shows settings-initiated download errors when a version is cached', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'error', message: 'ENOSPC' },
-        dismissedVersion: null,
-        cachedVersion: '1.2.0',
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('shows downloaded for card-initiated downloads', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'downloaded', version: '1.2.0' },
-        dismissedVersion: null,
-        cachedVersion: '1.2.0',
-        hasStartedDownload: true
-      })
-    ).toBe('visible')
-  })
-
-  it('re-shows card for a newer version even if an older version was dismissed', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'available', version: '1.3.0', changelog: null },
-        dismissedVersion: '1.2.0',
-        cachedVersion: '1.3.0',
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('shows error for dismissed version when an active update action fails', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'error', message: 'fail', userInitiated: true },
-        dismissedVersion: '1.2.0',
-        cachedVersion: '1.2.0',
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
-  })
-
-  it('hides check errors once a new checking cycle cleared the cached version', () => {
-    expect(
-      computeVisibility({
-        status: { state: 'error', message: 'network timeout' },
-        dismissedVersion: '1.2.0',
-        cachedVersion: null,
-        hasStartedDownload: false
-      })
-    ).toBe('hidden')
-  })
-})
-
-describe('HTTP/2 update error detection', () => {
-  it('recognizes Electron HTTP/2 protocol failures without matching generic errors', () => {
-    expect(isHttp2ProtocolError('net::ERR_HTTP2_PROTOCOL_ERROR')).toBe(true)
-    expect(isHttp2ProtocolError('Download failed: HTTP/2 protocol error')).toBe(true)
-    expect(isHttp2ProtocolError('Download failed: socket hang up')).toBe(false)
-    expect(isHttp2ProtocolError('HTTP proxy authentication failed')).toBe(false)
-  })
-})
-
-describe('manual update fallback for errors', () => {
-  it('suppresses manual download for signature-verification failures', () => {
-    expect(
-      shouldShowManualDownloadForUpdateError({
-        state: 'error',
-        message: 'Orca could not verify the update publisher.',
-        reason: 'signature-verification'
-      })
-    ).toBe(false)
-  })
-
-  it('keeps manual download available for ordinary update errors', () => {
-    expect(
-      shouldShowManualDownloadForUpdateError({
-        state: 'error',
-        message: 'download failed'
-      })
-    ).toBe(true)
-  })
-})
-
-describe('UpdateCard signature-verification error rendering', () => {
-  let root: Root | null = null
-  let container: HTMLDivElement | null = null
-
-  afterEach(() => {
-    root?.unmount()
-    root = null
-    container?.remove()
-    container = null
-  })
-
-  afterEach(() => {
-    useAppStore.setState({
-      updateStatus: { state: 'idle' },
-      updateChangelog: null,
-      updateUserInitiatedCycle: false,
-      dismissedUpdateVersion: null,
-      updateCardCollapsed: false
-    })
-  })
-
-  it('does not render the manual download fallback for signature failures', async () => {
-    useAppStore.setState({
-      updateStatus: {
-        state: 'error',
-        message: 'Orca could not verify the update publisher.',
-        reason: 'signature-verification',
-        userInitiated: true
-      },
-      updateUserInitiatedCycle: true,
-      updateCardCollapsed: false
-    })
-
-    container = document.createElement('div')
-    document.body.append(container)
-    root = createRoot(container)
-    await act(async () => {
-      root?.render(createElement(UpdateCard))
-    })
-
-    const html = container.innerHTML
-    expect(html).toContain('Update Verification Failed')
-    expect(html).toContain('Re-check')
-    expect(html).not.toContain('Download Manually')
-  })
-
-  it('renders 100 percent downloading as finalizing instead of stuck download copy', async () => {
-    useAppStore.setState({
-      updateStatus: { state: 'downloading', percent: 100, version: '1.2.0' },
-      updateUserInitiatedCycle: true,
-      updateCardCollapsed: false
-    })
-
-    container = document.createElement('div')
-    document.body.append(container)
-    root = createRoot(container)
-    await act(async () => {
-      root?.render(createElement(UpdateCard))
-    })
-
-    const html = container.innerHTML
-    expect(html).toContain('Orca v1.2.0 is being verified.')
-    expect(html).toContain('Finalizing update...')
-    expect(html).not.toContain('Downloading... 100%')
-  })
-})
-
 // ── Full update lifecycle through the store ──────────────────────────
 
 describe('full update lifecycle through setUpdateStatus', () => {
@@ -714,7 +346,7 @@ describe('full update lifecycle through setUpdateStatus', () => {
     expect(store.getState().updateChangelog).toBeNull()
   })
 
-  it('dismiss → new version cycle → card visible again', () => {
+  it('keeps dismissed update version scoped to the original version', () => {
     const store = createTestStore()
 
     setState(store, { state: 'available', version: '1.2.0', changelog: null })
@@ -725,14 +357,10 @@ describe('full update lifecycle through setUpdateStatus', () => {
     setState(store, { state: 'checking' })
     setState(store, { state: 'available', version: '1.3.0', changelog: null })
 
-    // The dismissed version is 1.2.0 but the available version is 1.3.0 — card should show
-    expect(
-      computeVisibility({
-        status: store.getState().updateStatus,
-        dismissedVersion: store.getState().dismissedUpdateVersion,
-        cachedVersion: '1.3.0',
-        hasStartedDownload: false
-      })
-    ).toBe('visible')
+    expect(store.getState().dismissedUpdateVersion).toBe('1.2.0')
+    expect(store.getState().updateStatus).toMatchObject({
+      state: 'available',
+      version: '1.3.0'
+    })
   })
 })
